nanoclaw/requirements.txt
# Discord
discord.py>=2.4.0

# LLM providers
anthropic>=0.42.0
openai>=1.50.0
google-generativeai>=0.8.0

# Config validation
pydantic>=2.10.0

# Database
aiosqlite>=0.20.0

# Git operations
GitPython>=3.1.44

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
